Added HQ logo
Pazaz sent me the cache with the updated HQ logo for the login screen so it no longer has artifacts/a pink pixel, I'm creating a new branch with this even though I don't know if this is the right location for the cache, oh well. I think this also contains missing models/midis.